Remove #if HAVE_X11 blocks from kstatusnotifieritem
They are all only around kwindowsystem framework usage which is not
needed to have in ifdefs. The runtime usage might differ, but from
looking at each line of code it works correctly. It's mostly about
virtual desktops and thus calls are simply ignored on systems not
supporting virtual desktops.

REVIEW: 115713