Avoid infinite recursion in KSNI
KStatusNotifierItem falls back to making a QSystemTray. With the KDE Qt
plugin our QSystemTrays create a StatusNotifierItem at which point we
are in infinite recursion.

This happens on bootup if we try to create a SNI before plasma is loaded but the KDE platform theme is being loaded. (for example the keyboard layout SNI which is in a KDED and loaded first)

REVIEW: 120449