Fix KSNIs being unable to register service under flatpak
Summary:
Rather than registering a custom service and then passing that to the
watcher's RegisterItem we can just pass the base service as a name.

As we're not registering random names it works on Flatpak.
Replaces D5911

Test Plan:
Ran the existing test with --ksni-count 5
Still had all my items appear

Reviewers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D5970